use usize as plainvoxelindex
- this is currently a limitation on the maximum number of voxels
  possible but we do not expect to exceed this at any reasonable time

+++ cellular_raza-concepts/src/domain_new.rs
@@ -49,7 +49,7 @@ where
     /// Entries are [Domain::SubDomainIndex], [SubDomain], and a vector of cells.
     pub index_subdomain_cells: Vec<(I, S, Vec<C>)>,
     /// Maps the assigned voxel index to the plain index indicated by a `u128` value.
-    pub voxel_index_to_plain_index: HashMap<S::VoxelIndex, u128>,
+    pub voxel_index_to_plain_index: HashMap<S::VoxelIndex, usize>,
     /// Encapsulates how the subdomains are linked to each other.
     /// Eg. two subdomains without any boundary will never appear in each others collection of neighbors.
     /// For the future, we might opt to change to an undirected graph rather than a hashmap.
